feat: upgrade to astro 5
