+++ package.json
@@ -23,13 +23,13 @@
     "toastify-js": "^1.12.0"
   },
   "devDependencies": {
-    "@astrojs/mdx": "^2.3.1",
-    "@astrojs/rss": "^4.0.9",
+    "@astrojs/mdx": "^4.0.3",
+    "@astrojs/rss": "^4.0.11",
     "@astrojs/sitemap": "^3.2.1",
-    "@astrojs/tailwind": "^5.1.2",
+    "@astrojs/tailwind": "^5.1.4",
     "@types/html-escaper": "^3.0.2",
     "@types/toastify-js": "^1.12.3",
-    "astro": "^4.16.11",
+    "astro": "^5.1.2",
     "hast-util-to-string": "^3.0.1",
     "hastscript": "^9.0.0",
     "html-escaper": "^3.0.3",
