+++ src/components/ProjectsFeaturedCard.astro
@@ -2,6 +2,7 @@
 import { Image } from "astro:assets"
 import RoundIcon from "./RoundIcon.astro"
 import { getIcon } from "../consts"
+import { render } from 'astro:content';
 
 interface Props {
   title: string
@@ -10,8 +11,8 @@ interface Props {
   heroImage: string
   stacks: string[]
   isFeatured?: boolean
-  index: number
-  render: any
+  index: number,
+  post: any
 }
 
 const {
@@ -22,10 +23,10 @@ const {
   index,
   demo,
   repo,
-  render
+  post
 } = Astro.props
 
-const { Content } = await render()
+const { Content } = await render(post)
 const iconStacks = stacks.filter((stack) => getIcon(stack)).reverse()
 ---
