--- src/content/config.ts
@@ -1,48 +1,0 @@
-import { defineCollection, z } from 'astro:content';
-
-const blog = defineCollection({
-	// Type-check frontmatter using a schema
-	schema: z.object({
-		title: z.string(),
-		description: z.string(),
-		// Transform string to Date object
-		pubDate: z
-			.string()
-			.or(z.date())
-			.transform((val) => new Date(val)),
-		updatedDate: z
-			.string()
-			.optional()
-			.transform((str) => (str ? new Date(str) : undefined)),
-		heroImage: z.string().optional(),
-    metaImage: z.string().optional(),
-    tags: z.array(z.string()),
-    otherUrl: z.string().optional(),
-    en: z.boolean().optional(),
-    draft: z.boolean().optional(),
-	}),
-});
-
-const project = defineCollection({
-	// Type-check frontmatter using a schema
-	schema: z.object({
-		title: z.string(),
-    repo: z.string().optional(),
-    demo: z.string().optional(),
-		// Transform string to Date object
-		pubDate: z
-			.string()
-			.or(z.date())
-			.transform((val) => new Date(val)),
-		updatedDate: z
-			.string()
-			.optional()
-			.transform((str) => (str ? new Date(str) : undefined)),
-		heroImage: z.string(),
-    isFeatured: z.boolean().optional(),
-    featuredOrder: z.number().optional(),
-    stacks: z.array(z.string()),
-	}),
-}); 
-
-export const collections = { blog, project };
