+++ src/layouts/BlogList.astro
@@ -1,5 +1,5 @@
 ---
-import { ViewTransitions } from "astro:transitions"
+import { ClientRouter } from "astro:transitions"
 import type { CollectionEntry } from "astro:content"
 import { Image } from "astro:assets"
 import BaseHead from "../components/BaseHead.astro"
@@ -22,7 +22,7 @@ const title = `${header} - ${SITE_TITLE}`
 <html lang="en">
   <head>
     <BaseHead title={title} description={SITE_DESCRIPTION} />
-    <ViewTransitions />
+    <ClientRouter />
   </head>
   <body>
     <Header />
@@ -38,13 +38,13 @@ const title = `${header} - ${SITE_TITLE}`
                     width={300}
                     height={150}
                     src={post.data.heroImage || ""}
-                    alt={post.slug}
+                    alt={post.id}
                     class=" w-full md:w-[300px] rounded-tr-lg md:rounded-tr-none rounded-tl-lg md:rounded-bl-lg h-full object-cover object-center"
                     loading="lazy"
                   />
                 </figure>
                 <div class=" w-full p-4 flex flex-col gap-1 border-l md:border-l-0 md:border-t border-b border-r border-gray-400 md:rounded-tr-lg rounded-br-lg rounded-bl-lg md:rounded-bl-none">
-                  <a href={`/blog/${post.slug}/`}>
+                  <a href={`/blog/${post.id}/`}>
                     <h4 class=" text-xl font-bold hover:underline underline-offset-4">
                       {post.data.title}
                     </h4>
