+++ src/pages/404.astro
@@ -1,5 +1,5 @@
 ---
-import { ViewTransitions } from "astro:transitions"
+import { ClientRouter } from "astro:transitions"
 import BaseHead from '../components/BaseHead.astro';
 import Header from '../components/Header.astro';
 import Footer from '../components/Footer.astro';
@@ -13,7 +13,7 @@ const title = '404 | ' + SITE_TITLE + ' - Frontend Developer'
 <html lang="en">
 	<head>
 		<BaseHead title={ title } description={SITE_DESCRIPTION} />
-    <ViewTransitions />
+    <ClientRouter />
 	</head>
 	<body class=" bg-gray-200 font-inter">
     <Header />
