+++ src/pages/blog/[...id].astro
@@ -1,18 +1,19 @@
 ---
 import { type CollectionEntry, getCollection } from 'astro:content';
+import { render } from 'astro:content';
 import BlogPost from '../../layouts/BlogPost.astro';
 
 export async function getStaticPaths() {
 	const posts = await getCollection('blog');
 	return posts.map((post) => ({
-		params: { slug: post.slug },
+		params: { id: post.id },
 		props: post,
 	}));
 }
 type Props = CollectionEntry<'blog'>;
 
 const post = Astro.props;
-const { Content, remarkPluginFrontmatter, headings } = await post.render();
+const { Content, remarkPluginFrontmatter, headings } = await render(post);
 ---
 
 <BlogPost {...post.data} remark={ remarkPluginFrontmatter } headings={ headings }>
