+++ src/pages/index.astro
@@ -1,5 +1,5 @@
 ---
-import { ViewTransitions } from "astro:transitions"
+import { ClientRouter } from "astro:transitions"
 import BaseHead from '../components/BaseHead.astro';
 import HomePage from '../layouts/HomePage.astro';
 import { SITE_TITLE, SITE_DESCRIPTION } from '../consts';
@@ -13,7 +13,7 @@ const title = SITE_TITLE + ' - Frontend Developer'
 <html lang="en">
 	<head>
 		<BaseHead title={ title } description={SITE_DESCRIPTION} />
-    <ViewTransitions />
+    <ClientRouter />
 	</head>
 	<body class=" bg-gray-200">
     <HomePage />
